requirements-dev.txt
# Local package
-e .

# External requirements
black
humanfriendly
jupyter
pandas
pocket
pylint
python-dotenv